+++ bpam/libs/excel_wrapper.py
@@ -18,7 +18,7 @@ import os
 import xlrd
 
 import logger_utils
- 
+
 logger = logger_utils.get_logger(__name__)
